refactor: mapActivity third/final part
- split mapActivity into smaller, focused components
- extracted location validation logic into seperate screens
- created dedicated controllersfor map and location handling
- improved state management for dialogs and map view
